+++ R/ace.R
@@ -165,7 +165,7 @@ ace_error <- function(ierr)
 ace.default  <- function(
   x,
   y,
-  wt      = rep(1, nrow(x)),
+  wt      = NULL,
   cat     = NULL, 
   mon     = NULL, 
   lin     = NULL,
@@ -177,6 +177,8 @@ ace.default  <- function(
   if(!is.null(control)) do.call(set_control, control)
   
   x  <- as.matrix(x)
+  
+  if(is.null(wt)) wt=rep(1, nrow(x))
   
   if (delrsq <= 0) stop("delrsq must be positive")